[Bug] Fixed progression switching is inconsistent with camera rails: switch character in the middle of rail transition (#70)
https://trello.com/c/HuRuz2rh/639-5-progression-system

Issus Summary:

In MainMenuWidget the SwitchCurrentPlayer was changing the new spot mesh component on the camera rail start which caused the PlayerTypeChanged event invoked faster than the actual camera rig transition completed. It was completely ignoring the camera rail state.

Solution:

I've introduced CameraRailTransitionState events. So now the MainMenuWidget sets the next Spot's mesh on the Camera rail transition on halfway of its path. The transition now from unlocked to locked level will be initiated in the middle of the camera rail.